Add stack erase checking and stack erase longjmp
A simple check that zeroes the first 2kb of stack before calling
main, then checks that it is all zeroes after main returns. The
value of the stack pointer is also checked after main returns - if
it is not where it was to begin with then that is an error.

The return codes are as follows:

SP value incorrect: 67
Stack not erased:   66
Otherwise:          <value returned from main>

A simple test program is as follows:

int main(void)
{
  int a;
  *(&a - 256) = 1;
  //__asm volatile ("addi sp, sp, 16\njr ra" ::);
  return 0;
}

This sets a value beyond the current frame on the stack so it does
not get erased, causing exit code 66.

To see an example of the SP value being incorrect resulting in
exit code 67, uncomment the __asm.

Longjmp also erases stack in this commit.

Both of these features are enabled using the --enable-stack-erase
configure flag.